+++ NEWS.md
@@ -1,5 +1,7 @@
-# tidyquant 0.3.0
+# tidyquant 0.2.0.9200
 
+* New data:
+    * New `tq_get` option `get = "key.stats"`, which retrieves the current key statistics (55 total) from [www.finance.yahoo.com/](http://finance.yahoo.com/). These include various current data such as Ask, Bid, Day's High, Day's Low, Last Trade Price, current P/E Ratio, EPS, Current Market Cap, EPS Projected Current Year, EPS Projected Next Year and many more. Example: `tq_get("AAPL", get = "key.stats")`.
 * New visualizations that integrate with `ggplot2`:
     * Chart geoms: Bar charts (`tq_geom_barchart`) and candlestick charts (`tq_geom_candlestick`) can be quickly created with the new geoms.
     * Moving Averages: Seven moving averages can be quickly visualized / prototyped using `tq_geom_ma`. The geom wraps the `TTR::SMA` functions.
@@ -14,7 +16,7 @@
 
 # tidyquant 0.2.0
 
-* Added `get = "key.ratios"` option for `tq_get()`, which retrieves 10-years of key performance ratios (89 total) from [www.morningstar.com](https://www.morningstar.com). These include various measures of financial performance including profitability, growth, cash flow, financial health, efficiency, and valuation ratios. Example: `tq_get("AAPL", get = "key.ratios")`.
+* New `get = "key.ratios"` option for `tq_get()`, which retrieves 10-years of key performance ratios (89 total) from [www.morningstar.com](https://www.morningstar.com). These include various historical measures of financial performance including profitability, growth, cash flow, financial health, efficiency, and valuation ratios. Example: `tq_get("AAPL", get = "key.ratios")`.
 * Added `zoo` `rollapply()` functions to list of compatible / integrated functions used with `tq_transform()` and `tq_mutate()`. See `tq_transform_fun_options()` for the full list.
 * Changed `tq_mutate()`, `tq_transform()`, `tq_mutate_xy()` and `tq_transform_xy()` arguments to be more obvious:
     * `x_fun` is now `ohlc_fun` for `tq_mutate()` and `tq_transform()`